+++ NEWS.rst
@@ -12,9 +12,14 @@ Unreleased
 ----------
 
 * (Put new changelog items here)
+* Allow for non-local addresses: for both listening interfaces and
+  connect endpoints, non-local addresses may be specified in a manner
+  similar to "ssh -L" or "ssh -R" arguments. See #37:
+  https://github.com/meejah/fowl/issues/37
 * Fix up some fallout from refactoring
 * Enable "remote" command in --interactive
 * Proper error-message rendering
+* Allow whitelisting only specific connect/listen endpoints.
 
 
 24.3.1: March 1, 2024